Add files via upload
Masterclass exercise 4. 
Finds whether the given year is a leap year or not.
If year is evenly divisible by 4, continue to next check. If not, return that it is not a leap year.
If evenly divisible by 100, continue to next check. Else return not a leap year.
If evenly divisible by 400, return it is a leap year. Else return it is not a leap year.