Add files via upload
Masterclass exercise 6. Takes 3 integer inputs. First two parameters are the values to be added to one another. The 3rd parameter is the number that the first two parameters should be equivalent to when added together. If the first two parameters added do not have a sum equivalent to the third parameter's value, returns false.